requirements.txt
boto3
google-cloud-bigquery
